// packages/ai-providers/shared-types.ts
export enum XyneTools {
  GetUserInfo = "get_user_info",
  MetadataRetrieval = "metadata_retrieval",
  Search = "search",
  FilteredSearch = "filtered_search",
  TimeSearch = "time_search",

  // Conversational tool
  Conversational = "conversational",

  // slack tools
  getSlackRelatedMessages = "get_slack_related_messages",
  getSlackThreads = "get_slack_threads",
  getUserSlackProfile = "get_user_slack_profile",
}